requirements.txt
PySide6-Fluent-Widgets>=1.1.0
requests>=2.25.0
